sparrow-research/requirements.txt
transformers==4.15.0
pytesseract==0.3.8
pyyaml==5.1
Pillow
numpy
datasets
seqeval